requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
together==1.3.11
